user model, authentication and authorization, bcrypt,  jwt, protected routes

+++ backend/controllers/goalControllers.js
@@ -1,8 +1,9 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
@@ -11,7 +12,7 @@ const createGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Please add a text field");
   }
-  const goal = await Goal.create({ text: req.body.text });
+  const goal = await Goal.create({ text: req.body.text, user: req.user.id });
   res.status(201).json(goal);
 });
 
@@ -20,6 +21,14 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal Not Found");
+  }
+  if (!req.user) {
+    res.status(401);
+    throw new Error("User Not Found");
+  }
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("Not Authorized");
   }
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
@@ -32,6 +41,14 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal Not Found");
+  }
+  if (!req.user) {
+    res.status(401);
+    throw new Error("User Not Found");
+  }
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("Not Authorized");
   }
   await goal.deleteOne();
   res.status(200).json({ id: req.params.id });
